msm: vidc: Fix broken debugfs creation error checks and error paths
The debugfs helper functions do not always return NULL when they fail;
instead, they can return an error number casted as a pointer, so that their
users have the option to determine the exact cause of failure.

Use the IS_ERR_OR_NULL() helper when checking for debugfs errors to fix the
error checks.

Change-Id: I937e294036482c13e1f40e5ea68db1673dcb6154
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: prorooter007 <shreyashwasnik112@gmail.com>
Signed-off-by: Poebe S.A <darkaccent69@gmail.com>